Fix some conversion warnings
https://bugzilla.gnome.org/show_bug.cgi?id=757295

+++ libsoup/soup-auth-manager.c
@@ -315,7 +315,7 @@ static char *
 soup_auth_manager_extract_challenge (const char *challenges, const char *scheme)
 {
 	GSList *items, *i, *next;
-	int schemelen = strlen (scheme);
+	gsize schemelen = strlen (scheme);
 	char *item;
 	GString *challenge;
